requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
feedparser==6.0.11
Flask==3.1.0
idna==3.10
importlib-metadata==6.8.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
python-dateutil==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.0
rfeed==1.1.1
redis==5.3.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
